update to offline functionality and api. now queue() returns an object that allows a user to register a callback to be executed once the request is made. several bugs were fixed as well.
more details:
queue() returns an object that contains a success, error, and then attribute. with these, a user will register a callback to be executed when the request is made and response is returned. within these
success callbacks, a user should remove the event handler if they plan on making the same request again, otherwise the callback will get executed twice.

the user may go offline and register requests. they will get stored and made when the device comes back online. I have provided a flush() method to clear out any pending requests when the app is loaded.
as long as event handlers are registered before flush() is called, then any requests in the queue will get flushed out and the appropriate handlers will get executed. flush() only works if the device is online. appworks-js
will automatically call flush() when the device goes from an offline state to an online state.